Changed the footer
The footer now includes the menu for people who have problems with the menu as it is at the top of the page.
Also changed the way the footer looks, it does no longer span the entire with of the page but stays with the width of all the other items on the page.