Changed some small astatic things
The index page now contains a header which is placed slightly over the slideshow and contains a red border on top. The icon for Nehalennia now scales up when the site becomes bigger and is also placed slightly overtop the slideshow. If this takes up too much space it should be removed.