Restructure project
Created a folder for the configs (config/) and put in the configs. Renamed the folder resources/img/ to resources/icon/. Downloaded the Nehalennia logo instead of linking to the source on the Nehalennia website. Put most includes in a new folder resources/includes; user-accessible pages are now in the project root. Renamed slideshow/header.php to slideshow/slideshow.php.