Fix wrong error message
Because it did only check the value of $_POST['dag'] it didn't work,
because we use 'vrijdag' and 'zaterdag'. The form works now but the
processing doesn't: it stops the script when it reaches the MySQL stuff.
This is likely the fault of wrong MySQL connecting data in config.php.
Maybe use MySQLite